feat(@angular-ru/common): move websocket to common
