+++ README.md
@@ -25,5 +25,4 @@ found in the Angular.
 | [@angular-ru/ng-excel-builder](https://npmjs.com/package/@angular-ru/ng-excel-builder)                 | ![](https://img.shields.io/npm/v/%40angular-ru%2Fng-excel-builder/latest.svg)         | [![](https://img.shields.io/badge/README--green.svg)](packages/ng-excel-builder/README.md)         | [![](https://img.shields.io/npm/dm/@angular-ru/ng-excel-builder)](https://npmjs.com/package/@angular-ru/ng-excel-builder)                 |
 | [@angular-ru/flex-layout](https://npmjs.com/package/@angular-ru/flex-layout)                           | ![](https://img.shields.io/npm/v/%40angular-ru%2Fflex-layout/latest.svg)              | [![](https://img.shields.io/badge/README--green.svg)](packages/flex-layout/README.md)              | [![](https://img.shields.io/npm/dm/@angular-ru/flex-layout)](https://npmjs.com/package/@angular-ru/flex-layout)                           |
 | [@angular-ru/stream](https://npmjs.com/package/@angular-ru/stream)                                     | ![](https://img.shields.io/npm/v/%40angular-ru%2Fstream/latest.svg)                   | [![](https://img.shields.io/badge/README--green.svg)](packages/stream/README.md)                   | [![](https://img.shields.io/npm/dm/@angular-ru/stream)](https://npmjs.com/package/@angular-ru/stream)                                     |
-| [@angular-ru/websocket](https://npmjs.com/package/@angular-ru/websocket)                               | ![](https://img.shields.io/npm/v/%40angular-ru%2Fwebsocket/latest.svg)                | [![](https://img.shields.io/badge/README--green.svg)](packages/websocket/README.md)                | [![](https://img.shields.io/npm/dm/@angular-ru/websocket)](https://npmjs.com/package/@angular-ru/websocket)                               |
 | [@angular-ru/eslint-plugin](https://npmjs.com/package/@angular-ru/eslint-plugin)                       | ![](https://img.shields.io/npm/v/%40angular-ru%2Feslint-plugin/latest.svg)            | [![](https://img.shields.io/badge/README--green.svg)](packages/eslint-plugin/README.md)            | [![](https://img.shields.io/npm/dm/@angular-ru/eslint-plugin)](https://npmjs.com/package/@angular-ru/eslint-plugin)                       |
