+++ packages/common/websocket/package.json
@@ -1,0 +1,8 @@
+{
+    "$schema": "../../../node_modules/ng-packagr/package.schema.json",
+    "ngPackage": {
+        "lib": {
+            "flatModuleFile": "angular-ru-common-websocket"
+        }
+    }
+}
