+++ packages/common/websocket/src/interfaces/websocket-handler.ts
