+++ packages/common/websocket/src/interfaces/websocket-message.ts
