+++ packages/common/websocket/src/websocket.module.ts
