--- packages/websocket/ng-package.json
@@ -1,10 +1,0 @@
-{
-    "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-    "lib": {
-        "entryFile": "./src/public_api.ts",
-        "umdModuleIds": {
-            "@angular-ru/common/typings": "angular-ru-common-typings"
-        }
-    },
-    "dest": "./dist/library"
-}
