--- packages/websocket/package.json
@@ -1,20 +1,0 @@
-{
-    "name": "@angular-ru/websocket",
-    "version": "15.304.0",
-    "description": "Angular-RU package",
-    "homepage": "https://github.com/Angular-RU/angular-ru-sdk/tree/master/packages/websocket#readme",
-    "bugs": "https://github.com/Angular-RU/angular-ru-sdk/issues",
-    "repository": "git+https://github.com/Angular-RU/angular-ru-sdk.git",
-    "license": "MIT",
-    "author": "splincode@yandex.ru",
-    "main": "lib/index.js",
-    "scripts": {
-        "build:lib": "ng build websocket-library && cp README.md dist/library",
-        "postversion": "npe version \"$(npe version)\" --package=\"$PWD/dist/library/package.json\"",
-        "test": "yarn jest --config ./jest.config.js"
-    },
-    "publishConfig": {
-        "access": "public",
-        "directory": "dist/library"
-    }
-}
