--- packages/websocket/tsconfig.base.json
@@ -1,12 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "rootDir": ".",
-        "baseUrl": ".",
-        "typeRoots": ["../../node_modules/@types", "./"],
-        "paths": {
-            "@angular-ru/websocket": ["src/public_api.ts"],
-            "@angular-ru/common/*": ["../common/*/src/public_api.ts"]
-        }
-    }
-}
