--- packages/websocket/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "extends": "./tsconfig.base.json",
-    "references": [
-        {
-            "path": "./tsconfig.lib.json"
-        },
-        {
-            "path": "./tsconfig.spec.json"
-        }
-    ]
-}
