--- packages/websocket/tsconfig.lib.json
@@ -1,9 +1,0 @@
-{
-    "extends": "./tsconfig.base.json",
-    "compilerOptions": {
-        "paths": {
-            "@angular-ru/common/*": ["../common/dist/library/*"]
-        }
-    },
-    "exclude": ["integration"]
-}
